Create HandleStaticVector. Change StaticVectors in allocators to HandleStaticVectors. Add fill to StaticVector
[git-p4: depot-paths = "//Asteroids-depot/": change = 49]

+++ src/Components/EntityAllocator.cpp
@@ -19,7 +19,7 @@ EntityAllocator* EntityAllocator::Get()
 
 int EntityAllocator::CreateEntity()
 {
-	unsigned int idx = entities.size();
+	unsigned int idx = (int)entities.size();
 	return entities.push_back({ idx });
 }
